1. Changed the function API prototype createJobAtInit() to have object. 2. Start JobServer after purging the redis DBs used by UI. 3. Pass orchModel in the jobServer wrapper functions to call proper auth APIs 4. If Auth is passed or there is no error while getting the response from API Server, then store authObj, tokenObjs and cookies in taskData in local DB. 5. For a recurrent JOB, once the job response comes to main server, do not delete the channel entry, as for recurrent job, for specific time interval, the job is rescheduled, and in case of redirectToLogout, we need to redirect to logout page, So we need to store that pubChannel entry. 6. Job Server jobs' taskData gets changed upon receiving the same job creating request. 7. When we get auth failure from API Server, and we do redirect logout, then do redirect to logout only for that session.

+++ src/serverroot/web/core/cache.api.js
@@ -345,21 +345,22 @@ function createReqData (req, type, jobName, reqUrl, runCount, defCallback,
  This function is used to insert the response published on redis channel on
  ready Q
  */
-function sendResponseByChannel (channel, msg)
+function sendResponseByChannel (channel, msg, deleteChannelAtMainServer)
 {
     var pendClientLists = checkCachePendingQueue(channel);
     if (null == pendClientLists) {
         return;
     }
     var isJson = 0;
-    var msgParse = JSON.parse(msg);
-    if (global.HTTP_STATUS_RESP_OK == msgParse.errCode) {
+    if (global.HTTP_STATUS_RESP_OK == msg.errCode) {
         isJson = 1;
         /* In error case, application will send only the error string */
     }
-    deleteCachePendingQueueEntry(channel);
-    longPoll.insertDataToSendAllClients(pendClientLists, msgParse.data, 
-                                        msgParse.errCode, isJson);
+    if (true == deleteChannelAtMainServer) {
+        deleteCachePendingQueueEntry(channel);
+    }
+    longPoll.insertDataToSendAllClients(pendClientLists, msg.data,
+                                        msg.errCode, isJson);
 }
 
 function handleJSONResponse(error, req, res, jsonStr)
